Reduce the number of retries in temp_file and open_temp_file to 20.
Historically, the number was high because temporary file names were
assigned in a fixed order.  But now, filenames are chosen randomly, so
the odds of a collision are much smaller.  We would prefer to fail
fast in the event of some filesystem problem that prevents file creation
(for instance, a missing temp_dir, or running out of inodes).