Pack gc_regs buckets in arm64.S as in amd64.S
Each bucket in gc_regs_buckets uses the first value to store a pointer to
the next bucket in the list. In amd64.S, this slot is then used for the
registers, but in arm64.S a slight dance was done to write a pointer to
the _third_ index of the bucket instead.

Implement SAVE_ALL_REGS in arm64.S in the same way as amd64.S, saving two
words in the gc_regs bucket for arm64.