Only write gc_regs for caml_call_gc
In amd64.S, SAVE_ALL_REGS returns the gc_regs bucket in %r15. This
leaves it on the caller to write it to the gc_regs field of the domain
state which is done (obviously) for caml_call_gc, but is not done for
caml_call_realloc_stack.

Alter the arm64.S implementation to use x19 in the same way.