[refactoring] runtime/frame_descriptors: store num_descr in the table
To solve quadratic-slowdown issues with the current frame_descriptors
code, we want to go back to the 4.14 logic where frame-descriptor
fragments grow an existing hash table, instead of rebuilding
a hashtable from scratch everytime.

This is easier to implement clearly if the hashtable carries the
number of descriptors currently in (to know when to resize). We also
need the capacity, but we can recompute it from the "mask" already stored.