[refactoring] frame_descriptors.c: add `add_frame_descriptors`
This new function adds a new list of frame descriptors to an existing
list and hashtable. This feature is not currently used in the code,
where we only ever add frame descriptors to the empty list -- but it
will be necessary to fix the quadratic behavior of repeated calls to
`caml_register_frametables`.

The new code is in fact a direct port of the 4.14 implementation
(the weirdly-named [init_frame_descriptors] in roots_nat.c). Reviewers
could gain confidence in the code by comparing to that 4.14
implementation -- the only non-mechanical change is a use of
[caml_stat_calloc]  instead of explicit NULL-initialization.