Implement `ocamlc -g1` (generation of lightweight debug info)
Just remove all typing informations and informations about local
variables from the debugging events.  This does not affect the
generation of stack backtraces, and some debugging with ocamldebug is
still possible (but no printing of values, of course).

The `-g1` option takes effect when compiling to a .cmo but also when
linking to a bytecode executable.  In particular, it is possible to
compile a library with `-g`/`-g2` to get full debugging info, yet link
with `-g1`, if smaller debugging info is desired.