runtime: move last finalisers book keeping during a minor cycle to the end of the minor cycle
This commit implements a fix on how last finalisers are updated during the minor cycle.
Previously, the book keeping was done during the minor cycle, before the values are promoted.
As a result, last finalisers are wrongly run after the end of the minor cycle, precisely because
the book keeping did not account the the promoted values.
The fix is to move said book keeping to the end of the minor gc cycle, and can be done
outside of the last barrier for the minor cycle, as a domain will only process finalisers
within its own minor heap.

Fixes #11995