Misc.find_first_mono
Several places in the compiler have logic to "pick a fresh name"
by starting from a prefix, say foo, and trying foo/0, foo/1,
foo/2... until they find one that is "available" in some sense.

This can easily be implemented through a linear search, but this gives
a quadratic behavior when we repeatedly generate fresh names from the
same prefix -- which can happen in practice in realistic example of
source code repeating the same variable or constructor name many
times.

We introduce a function that finds the "first available number" in
logarithmic time, assuming that the "unavailable numbers" form
a strict prefix of all natural numbers. (That is: some numbers remain
available, and all numbers after the first available numbers are
available.)

When called repeatedly on the same prefix, this gives a behavior in
O(n log n) rather than O(n^2).

Note: if called on availability predicates that are not
monotonous (they become false again after being true), the function
will not necessarily provide the smallest available number, but it
should still provide an available number.