intext_par: avoid a sequentiality assumption on object identifiers (#12069)
In presence of concurrent code execution, we cannot assume that two
object creation in sequence will return consecutive integers.

(The testcase was added as a regression test for a case where object
identifiers were all constant after demarshalling. The weaker test
that we propose would also catch this issue if it came back.)

Note: due to the batching of object identifiers, it is in fact very
hard to observe a failure of this test: each domain reserves the next
1024 identifiers, so in most case we observe consecutive identifiers
for the same domain.