Add a new debug runtime parameter to efficiently trace the current instruction count
This adds CAMLRUNPARAM=C=<num> where num is the number of instructions the
debug runtime will wait before printing the current instruction counter.

If the trace-level (CAMLRUNPARAM=t=2 or above) is present, this new parameter
will be ignored and the instruction counter will be printed for each
instructions as before.