Make private build variables private
Given the convention that the OC_* build varialbes are reserved for
the build system, it seems better to make sure all of them are defined
in the private Makefile.build_config file, rather than in Makefile.config
which gets installed and thus becomes public.

This commit moves the definitions of OC_CFLAGS, OC_CPPFLAGS,
OC_LDFLAGS, OC_DLL_LDFLAGS and OC_EXE_LDFLAGS from Makefile.config.in to
Makefile.build_config.in. It also moves the defintion of MKEXE_VIA_CC,
since this variable relies on private build varables and does not seem
relevant or useful outside of the context of the build of the compiler itself.