Refactor to remove `type_unpacks`
This is a refactor to avoid the need to wrap typechecking the
extent of a module unpack in the `type_unpacks` function.
Now you call `add_module_variables` to add the (typechecked)
module bindings to the environment before calling `type_expect`
on the extent; this is more similar to how we add variable bindings
to the environment.