Add deployment phase support for BUILD_PATH_PREFIX_MAP
The BUILD_PATH_PREFIX_MAP specification tells how to use that
environment variable to achieve reproducible build, i.e. builds of
products that do not leak absolute paths from the build
environment. See
https://reproducible-builds.org/specs/build-path-prefix-map.

However, that specification only describes half of the story.  Let us
call the building of reproducible products the "Build Phase". That is
the phase covered by the existing specification.

Let us defined the "Deployment phase" as the phase where you accept a
built reproducible product and make use of it, i.e. deploy it. An
example would be the debugger taking a reproducible binary and letting
the user debug it, showing the user the source code, etc.

We use the same mechanism in the deployment phase, but use a different
variable, DEPLOY_PATH_PREFIX_MAP.  It must be setup with the logical
inverse of the mapping used during the build phase.

Also, this is part of a larger PR,
https://github.com/ocaml/ocaml/pull/12085, that is being split up
because it was too large. In addition to these changes, that PR
includes ocamltest enhancements plus the tests for these changes. See
it for prior discussion and the other changes.

I will now describe the essence of the changes.

1. bytecomp/emitcode.ml

Added sanitizing of paths in the debug events.
If the compiler is given absolute source paths
this was leaking absolute build paths.

2. debugger/command_line.ml

A debugger variable, "mapping", was added, which is tied to the
DEPLOY_PATH_PREFIX_MAP environment variable.  So "set mapping value"
will set DEPLOY_PATH_PREFIX_MAP, and "show mapping" will show the value
of DEPLOY_PATH_PREFIX_MAP. This allows the user to set it
programmaticaly, possibly from the ".ocamldebug" file.

3. debugger/source.ml

In source_of_module, use the Location.rewrite_exists.

4. debugger/symbols.{ml, mli}

a. Avoid adding directories redundantly

b. Use Location.matching_dirs to expand directories.

5. Man page and manual updated.

Updated new tests to take into account:

1. The new ocamltest syntax that is C-like rather than
org-mode like.

2. The debugger now uses DEPLOY_PATH_PREFIX_MAP for its
inverse mapping, rather than BUILD_PATH_PREFIX_MAP.

Fixes #12083