Explain type errors arising from string indexing
Joint work with @spdskatr at the Tarides Compiler Hacking Day.

Fixes #11543 by adding an explanation to the error message:

    # ListLabels.map ~f:(fun i -> i+1). [1;2;3];;
    Error: This expression should not be a function, the expected type is
           string because it is in a string-indexing expression

This may not fully explain the error message in this case (in particular the
surprising lexing behavior) but at least the user has something to Google.

Also applies to array indexing (both single- and multi-dimensional).