Add a -build-tree option to ocamldep
ocamldep generates the list of dependencies of a given OCaml target
file. The new -build-tree option adds a facility to add a (separate)
prefix to the targets and the dependencies in its make output. This
allows make to write targets in an other location than just the
current working directory. Dependencies of the target that are also
build artifacts are also found in this location.

This is useful for out-of-source builds and cross-compilation.

ocamldep outputs a trailing directory separator if the given string
doesn't contain one, and respects -slash and -no-slash options. No
other processing is made on the string. It is for instance possible to
give a Makefile variable instead of a path.