bind anonymous rows for polyvariants to nearest binder
Anonymous rows in polyvariant types like ``[> `Foo]`` are generally
bound at the global variable levels. This is fine in absence of
explicitly polymorphic type annotation.

However, as soon that such an anonymous row appears under such an
annotation, the row variable might capture the explicitly
polymorphic variants leading to an error in OCaml 5.1 due to a bug
fix

```
let f: 'a. [> `Foo of 'a ] -> unit = fun _ -> ()
```
```
Error: This pattern matches values of type [> `Foo of 'a ]
       but a pattern was expected which matches values of type 'b
       The universal variable 'a would escape its scope
```

To decrease the pain created by this bug fix, this commit binds
anonymous row variables of syntactic polyvariants to the nearest
binder:

```
let f: 'a. [> `Foo of 'a ] -> unit = fun _ -> ()
```

is interpreted as

```
let f: 'a 'r. ([> `Foo of 'a ] as 'r) -> unit = fun _ -> ()
```

This has the advantage of preserving the ability to use the shorthand
`type a b c. ... ` when mixing GADTs and polymorphic variants in OCaml
5.1

```
type 'a w = Int: int w
let f: type a. a w -> [> `Foo of a ] -> unit = fun Int ->
function `Foo x -> x | _ -> 0
```
is elaborated into
```
let f: 'a. 'a w -> [> `Foo of 'a ] -> 'a = (fun (type a) -> ... )
```
which is correct with the new rule in this commit that the anonymous
row variable in `[> Foo ... ]` is bound at the `'a. ... ` level.

However, beware this commit is a breaking change: the user needs to add
an `as` annotation if they really intended to use an unification
variable:

```ocaml
let f: 'b. [< `Foo | `Bar ] -> 'b -> 'b = fun `Foo x -> x
```
needs to be written after this commit

```ocaml
let f: 'b. ([< `Foo | `Bar ] as 'r) -> 'b -> 'b = fun `Foo x -> x
```

Fix #12189

Co-authored-by: Florian Angeletti <florian.angeletti@inria.fr>