Make the OCAML_{BYTECODE,NATIVE}_PROGRAM macros more robust
So far, program targets could only accept object files to be linked
into the built program as their prerequisites. This was so because the recipes
were using $^ as the list of object files to link, meaning all
the prerequisites.

With this commit, it becomes possible to have prerequisites that
do need to be build before the target is built but without necessarily
being linked into the target.

This is implemented by defining target-specific variables foo_BCOBJS
and foo_NCOBJS containing the list of associated object files.

These variables are then used in the recipes, meaning that the
targets can now take additional prerequisites which won't appear in the recipes.