ocamltest/Makefile: use wildcards when computing dependencies
This is to bring the computation of dependencies closer to what the root
Makefile will do. Sadly, we use wildcards rather than precise lists of
files because it is what is going to happen in the root Makefile, at
least (hopefully only) for the time being.

This adds the dependencies of the two implementations of the
ocamltest_unix module which are not necessary but don't hurt either.

Apart from that, although this commit modifies the order in which
the files are passed to ocamldep, the generated .depend remains untouched
because ocamldep prints the dependencies in sorted order.

In other words, the order in which files are passed to ocamldep
does not affect the order in which dependencies are printed.