Fix infinite loop in signal handling.
Co-authored-by: Guillaume Munch-Maccagnoni <Guillaume.Munch-Maccagnoniමinria.fr>

This commit fixes a race in the signal handling code that creates an
infinite loop in `caml_enter_blocking_section`.

In order to process signals, the function
`caml_process_pending_signals_exn` first checks whether signals are
pending in the flag (`caml_pending_signals` array) and whether the
signal is part of the current signal set. Then, it clears the pending
status in the flag for that signal and calls `caml_execute_signals_exn`,
which masks the signal in the current thread's signal set while the
OCaml handler is running.

It is possible that a signal may arrive after the signal pending flag is
reset but before the signal is masked. This puts the system in
inconsistent state.

The signal handler may call `caml_enter_blocking_section. The function
`caml_enter_blocking_section` ensures that all the pending signals are
processed before entering the blocking section by calling
`caml_process_pending_signals_exn` **until the pending flag is
cleared**.

Recall that the signal is masked in the thread's signal set. In order to
process the pending signal, `caml_process_pending_signals_exn` checks
whether signals are pending by checking the flag (yes) and also whether
the signal is part of the current signal set (no). So the signal is not
processed. Hence, `caml_enter_blocking_section` goes into an infinite
loop.

The fix here is to perform a weaker check for pending signals by
checking whether the `young_limit` is set to `UINTNAT_MAX` (which is the
mechanism by which a domain is interrupted when a signal arrives, but
also in other situations).