manual: explain the difference between multi-argument and single-tuple-argument variant constructors
fixes #7783, see also
https://discuss.ocaml.org/t/sum-type-constructor-declaration-subtlety-and-documentation-wanted/1898

The new content is included at the end of the (fairly brief)
presentation of records and variants in the "tutorial / language tour"
part of the manual.

This is arguably more advanced content and it could also somewhere
else, but it is unclear that the "reference manual" part of the manual
would be more appropriate.

It depends on how we think of this explanation/warning and its
intended target audience: is it a pitfall that all OCaml beginners
should be told about to avoid confusion later (similar to: `Array.make
xlen (Array.make ylen 0)`), or is it advanced content that we want
teachers/mentors to be able to point to as a reference explanation?

This commits tries to aim for the "beginner tour" aspect, as suggested
by past discussions, so it is intentionally brief and light on
details.