Reduce size of installation bindir by ~20MB
For reasons of safety in _user_ programs, ocamlcommon.cma and
ocamlcommon.cmxa set -linkall. This is not necessary for the
distributions programs themselves, and makes the size of various tools
(ocamldep, etc.) considerably larger than they need to be.

The compiler's build now uses ocamlcommon-private, which is linked
without -linkall, but it installs the same ocamlcommon as before. The
compilers use all the modules in ocamlcommon, so this results in
identical compilers, but smaller tools.