Make Unix.create_process_env thread-safe in all cases
If neither `spawnp` nor `execvpe` are provided, the emulation code ended up allocating memory in the child process using `malloc`, which is not thread-safe.

As a fix, in the child process we just set `environ` and call `execvp`.  This is safe because no other thread is running in the child process.