remove Location.lines_around_from_file -- #12238
This logic is not robust in presence of lexer directives and will
silently do the wrong file. We don't need it in the compiler anymore
now that we read the entire file at once -- we do not need a fallback
strategy after lines_around_from_lexbuf anymore.

In theory this might make a difference for compiler-libs users that
would set Location.input_name but not Location.input_lexbuf, and rely
on the read-from-file fallback logic. Those users can fix their
code (in a backward-compatible way) by setting Location.input_lexbuf
themselves.