Report the same BSD systems for x86_32 and x86_64
The 64-bit ports of OCaml have `Config.system` being freebsd, netbsd or
openbsd, as appropriate but the 32-bit Intel ports have bsd for all of
them. The x86_32 now reports the same value as for x86_64, which in
particular means that BSDs matching *bsd but not {free,net,open}bsd are
no longer identified.