Workaround GNU make 4.4 new warning
GNU make 4.4.1 is triggering an undefined variable warning on the use of
'$(BOOT_OCAMLC)' in a definition in the command line. It's not clear if
this is intentional or not, but as the whole mechanism will ultimately
disappear when stdlib/Makefile is merged into the root Makefile, this
provides a viable workaround for the present setup.