fix #12494: open shadowing warning mistriggers
The environment contains some "dummy" entries with constructor
Mod_unbound and Val_unbound. They should not count as a "previous
occurrence" of a name shadowed by an "open" statement.

The relevant function to understand this change are
`IdTbl.find_name` (and `open_signature`, not shown):

    let rec find_name wrap ~mark name tbl =
      try
        let (id, desc) = Ident.find_name name tbl.current in
        Pident id, desc
      with Not_found as exn ->
        begin match tbl.layer with
        | Open {using; root; next; components} ->
            begin try
              let descr = wrap (NameMap.find name components) in
              let res = Pdot (root, name), descr in
              if mark then begin match using with
              | None -> ()
              | Some f -> begin
                  match find_name wrap ~mark:false name next with
                  | exception Not_found -> f name None
                  | _, descr' -> f name (Some (descr', descr))
                end
              end;
              res
            with Not_found ->
              find_name wrap ~mark name next
            end
        | Map {f; next} ->
            let (p, desc) =  find_name wrap ~mark name next in
            p, f desc
        | Nothing ->
            raise exn
        end

When we use a value from an `open` layer, and `mark` holds, we call
the `using` function with metadata about the previous binding of the
same name in the environment, or None if the name was not previously
bound. The `using` function corresponds to the function that was
passed as `slot` to `open_signature`.

The `check_shadowing` is called by the `using`
function (see `open_signature`) to decide whether to emit a shadowing
warning or not.

The effect of this commit is thus that, when the previous binding in
the environment is an "unbound" binding, the shadowing warning is
ignored. It is not obvious that this is the right semantics, one may
have expected that unbound bindings would be ignored, and that
`check_shadowing` would be called with the binding before that in the
environment. But in fact disabling the warning altogether (instead of
ignoring unbound bindings) is the right choice:

- Most `{Mod,Val}_unbound` bindings are meant to prevent using the
  name completely in the unbound environment, for example by making
  recursive references to the module illegal. The user cannot use
  the name anymore (in particular, they do not get the "previous"
  binding in the environment), so shadowing does not introduce
  any ambiguity.

- The `Val_unbound` constructor used for the
  possibly-recursive-definition hint (involved in the bug report) is not
  meant to cancel/hide previous bindings, but it is only introduced in
  the environment if there are no previous bindings in the environment,
  so disabling the warning (instead of looking for previous bindings0 is
  also correct.