Use strict prototypes on primitives in bytecomp
The bytecode compiler also extracts the list of primitives to a C
file, similarly as the Makefile. The Makefile was fixed in
8db1817ac0e9fde774c656527e5609950dc2ee79 to account for stricter C
function prototypes, but the bytecomp problems were not discovered
back then.

As the startup code includes `caml/mlvalues.h`, two primitives are
declared twice with conflicting types. We prevent this with
preprocessor tricks before including the header.

The generated C code now also uses the same names as in the Makefile.